[x86] spec_ctrl: Add casting to fix compilation error
Message-id: <20190514171515.7752-4-longman@redhat.com>
Patchwork-id: 255203
O-Subject: [RHEL7.7 PATCH 03/42] x86/spec_ctrl: Add casting to fix compilation error
Bugzilla: 1709296 1690358 1690348 1690335
CVE: CVE-2018-12126 CVE-2018-12127 CVE-2018-12130 CVE-2019-11091
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>
RH-Acked-by: Phil Auld <pauld@redhat.com>

Bugzilla: 1690335 1690348 1690358
Upstream: RHEL only

In an upcoming patch, all the SPEC_CTRL_* macros will be converted to
unsigned long by way of the BIT() macro. This may cause compilation
error when the negated forms are used. So adding casting to all those
cases to avoid it.

+++ arch/x86/kernel/spec_ctrl.c
@@ -661,8 +661,8 @@ void spec_ctrl_set_ssbd(bool ssbd_on)
 		this_cpu_or(spec_ctrl_pcp.entry, SPEC_CTRL_SSBD);
 		this_cpu_or(spec_ctrl_pcp.exit,  SPEC_CTRL_SSBD);
 	} else {
-		this_cpu_and(spec_ctrl_pcp.entry, ~SPEC_CTRL_SSBD);
-		this_cpu_and(spec_ctrl_pcp.exit,  ~SPEC_CTRL_SSBD);
+		this_cpu_and(spec_ctrl_pcp.entry, ~(int)SPEC_CTRL_SSBD);
+		this_cpu_and(spec_ctrl_pcp.exit,  ~(int)SPEC_CTRL_SSBD);
 	}
 }
